Move the common test tools into their own package (#40)
