+++ core-android/src/test/java/com/mux/core_android/test/tools/Log.kt
@@ -1,6 +1,4 @@
-package com.mux.core_android
-
-import java.lang.Exception
+package com.mux.core_android.test.tools
 
 fun log(tag: String = "\t", message: String, ex: Throwable? = null) {
   println("$tag :: $message")
